Renamed client_app_<platform> to appshell_helpers_<platform>
