+++ appshell/appshell_extensions_gtk.cpp
@@ -23,7 +23,8 @@
 
 #include "appshell_extensions_platform.h"
 
-#include "client_app.h"
+#include "appshell/appshell_helpers.h"
+
 #include <glib.h>
 #include <glib/gstdio.h>
 #include <gio/gio.h>
@@ -70,7 +71,7 @@ int32 OpenLiveBrowser(ExtensionString argURL, bool enableRemoteDebugging)
     GError *gerror = NULL;
     
     if (enableRemoteDebugging) {
-        CefString appSupportDirectory = ClientApp::AppGetSupportDirectory();
+        CefString appSupportDirectory = appshell::AppGetSupportDirectory();
 
         // TODO: (INGO) to better understand to string conversion issue, I need a consultant
         // here. Getting the char* from CefString I had to call ToString().c_str()
