+++ appshell/appshell_extensions_win.cpp
@@ -23,9 +23,8 @@
 
 #include "appshell_extensions_platform.h"
 
-#include "client_app.h"
+#include "appshell/appshell_helpers.h"
 #include "native_menu_model.h"
-#include "client_handler.h"
 
 #include <algorithm>
 #include <CommDlg.h>
@@ -339,7 +338,7 @@ int32 OpenLiveBrowser(ExtensionString argURL, bool enableRemoteDebugging)
     std::wstring args = appPath;
 
     if (enableRemoteDebugging) {
-        std::wstring profilePath(ClientApp::AppGetSupportDirectory());
+        std::wstring profilePath(appshell::AppGetSupportDirectory());
         profilePath += L"\\live-dev-profile";
         args += L" --user-data-dir=\"";
         args += profilePath;
