+++ appshell/appshell_helpers_mac.mm
@@ -21,24 +21,24 @@
  *
  */
 
-#include "client_app.h"
+#include "appshell/appshell_helpers.h"
 
 #include "include/cef_base.h"
 #include "config.h"
 #include <Cocoa/Cocoa.h>
 
-#include <string>
-
 extern CFTimeInterval g_appStartupTime;
 
-double ClientApp::GetElapsedMilliseconds()
+namespace appshell {
+
+double GetElapsedMilliseconds()
 {
     CFAbsoluteTime elapsed = CFAbsoluteTimeGetCurrent() - g_appStartupTime;
-    
+
     return round(elapsed * 1000);
 }
 
-CefString ClientApp::GetCurrentLanguage()
+CefString GetCurrentLanguage()
 {
     // Do not confuse preferredLanguages with currentLocale.
     // preferredLanguages specifies to UI language. currentLocale
@@ -49,24 +49,24 @@ CefString ClientApp::GetCurrentLanguage()
     // default on US systems.
     // NSLog(@"localeIdentifier: %@", [[NSLocale currentLocale] localeIdentifier]);
     // NSLog(@"%@", language);
-  
+
     NSString* language = [[NSLocale preferredLanguages] objectAtIndex:0];
-    
+
     CefString result = [language UTF8String];
     return result;
 }
 
-std::string ClientApp::GetExtensionJSSource()
+std::string GetExtensionJSSource()
 {
     std::string result;
-    
+
     // This is a hack to grab the extension file from the main app resource bundle.
     // This code may be run in a sub process in an app that is bundled inside the main app.
 #if 1
     NSString* bundlePath = [[NSBundle mainBundle] bundlePath];
     NSString* sourcePath;
     NSRange range = [bundlePath rangeOfString: @"/Frameworks/"];
-    
+
     if (range.location == NSNotFound) {
         sourcePath = [[NSBundle mainBundle] pathForResource: @"appshell_extensions" ofType: @"js"];
     } else {
@@ -76,23 +76,30 @@ std::string ClientApp::GetExtensionJSSource()
 #else
     NSString* sourcePath = [[NSBundle mainBundle] pathForResource: @"appshell_extensions" ofType: @"js"];
 #endif
-    
+
     NSString* jsSource = [[NSString alloc] initWithContentsOfFile:sourcePath encoding:NSUTF8StringEncoding error:nil];
     result = [jsSource UTF8String];
     [jsSource release];
-    
+
     return result;
 }
 
+CefString AppGetSupportDirectory() {
+    NSString *libraryDirectory = [NSSearchPathForDirectoriesInDomains(NSApplicationSupportDirectory, NSUserDomainMask, YES) objectAtIndex:0];
+    NSString *supportDirectory = [NSString stringWithFormat:@"%@/%@%@", libraryDirectory, GROUP_NAME, APP_NAME];
 
-CefString ClientApp::AppGetSupportDirectory() {
-  NSString *libraryDirectory = [NSSearchPathForDirectoriesInDomains(NSApplicationSupportDirectory, NSUserDomainMask, YES) objectAtIndex:0];
-  NSString *supportDirectory = [NSString stringWithFormat:@"%@/%@%@", libraryDirectory, GROUP_NAME, APP_NAME];
-  
-  return CefString([supportDirectory UTF8String]);
+    return CefString([supportDirectory UTF8String]);
 }
 
-CefString ClientApp::AppGetDocumentsDirectory() {
+CefString AppGetDocumentsDirectory() {
     NSString *documentsDirectory = [NSSearchPathForDirectoriesInDomains(NSDocumentDirectory, NSUserDomainMask, YES) objectAtIndex:0];
     return CefString([documentsDirectory UTF8String]);
 }
+
+CefString AppGetCachePath() {
+    std::string cachePath = std::string(AppGetSupportDirectory()) + "/cef_data";
+
+    return CefString(cachePath);
+}
+
+}  // namespace appshell
