+++ appshell/cefclient_mac.mm
@@ -18,6 +18,7 @@
 #include "config.h"
 #include "command_callbacks.h"
 #include "appshell/common/client_switches.h"
+#include "appshell/appshell_helpers.h"
 #include "native_menu_model.h"
 #include "appshell_node_process.h"
 
@@ -840,7 +841,7 @@ int main(int argc, char* argv[]) {
     
   // Check command
   if (CefString(&settings.cache_path).length() == 0) {
-	  CefString(&settings.cache_path) = AppGetCachePath();
+	  CefString(&settings.cache_path) = appshell::AppGetCachePath();
   }
 
   // Initialize CEF.
@@ -918,12 +919,6 @@ int main(int argc, char* argv[]) {
 
 std::string AppGetWorkingDirectory() {
   return szWorkingDir;
-}
-
-CefString AppGetCachePath() {
-  std::string cachePath = std::string(ClientApp::AppGetSupportDirectory()) + "/cef_data";
-  
-  return CefString(cachePath);
 }
 
 CefString AppGetProductVersionString() {
