+++ appshell/cefclient_win.cpp
@@ -18,6 +18,7 @@
 #include "config.h"
 #include "appshell/browser/resource.h"
 #include "appshell/common/client_switches.h"
+#include "appshell/appshell_helpers.h"
 #include "native_menu_model.h"
 #include "appshell_node_process.h"
 
@@ -232,7 +233,7 @@ int APIENTRY wWinMain(HINSTANCE hInstance,
 
   // Check command
   if (CefString(&settings.cache_path).length() == 0) {
-	  CefString(&settings.cache_path) = AppGetCachePath();
+	  CefString(&settings.cache_path) = appshell::AppGetCachePath();
   }
 
   // Initialize CEF.
@@ -352,13 +353,6 @@ BOOL InitInstance(HINSTANCE hInstance, int nCmdShow) {
 
 std::string AppGetWorkingDirectory() {
   return szWorkingDir;
-}
-
-CefString AppGetCachePath() {
-  std::wstring cachePath = ClientApp::AppGetSupportDirectory();
-  cachePath +=  L"/cef_data";
-
-  return CefString(cachePath);
 }
 
 CefString AppGetInitialURL() {
