+++ appshell_paths.gypi
@@ -182,6 +182,7 @@
       'appshell/appshell_extensions.h',
       'appshell/appshell_extensions_platform.h',
       'appshell/appshell_extensions.js',
+      'appshell/appshell_helpers.h',
       'appshell/appshell_node_process.h',
       'appshell/appshell_node_process_internal.h',
       'appshell/appshell_node_process.cpp',
@@ -225,7 +226,7 @@
       'appshell/cefclient.rc',
       'appshell/version.rc',
       'appshell/cefclient_win.cpp',
-      'appshell/client_app_win.cpp',
+      'appshell/appshell_helpers_win.cpp',
       'appshell/client_handler_win.cpp',
       'appshell/res/appshell.ico',
       'appshell/cef_buffered_dc.h',
@@ -280,7 +281,7 @@
       'appshell/FullScreenViewController.mm',     
       'appshell/appshell_extensions_mac.mm',
       'appshell/appshell_node_process_mac.mm',
-      'appshell/client_app_mac.mm',
+      'appshell/appshell_helpers_mac.mm',
       'appshell/cefclient_mac.mm',
       'appshell/client_handler_mac.mm',
       '<@(appshell_sources_browser)',
@@ -303,7 +304,7 @@
       'appshell/FullScreenViewController.mm',
       'appshell/appshell_extensions_mac.mm',
       'appshell/appshell_node_process_mac.mm',
-      'appshell/client_app_mac.mm',
+      'appshell/appshell_helpers_mac.mm',
       'appshell/client_handler_mac.mm',
       'appshell/process_helper_mac.cpp',
       '<@(appshell_sources_common_helper)',
@@ -383,7 +384,7 @@
 
       'appshell/appshell_extensions_gtk.cpp',
       'appshell/appshell_node_process_linux.cpp',
-      'appshell/client_app_gtk.cpp',
+      'appshell/appshell_helpers_gtk.cpp',
       'appshell/client_handler_gtk.cpp',
 
       '<@(appshell_sources_browser)',
